extract general modal
